Update simulation to use values from db

+++ server/migrations/V4_dynamic_pricing.js
@@ -1,8 +1,8 @@
 export async function up(knex) {
   await knex.raw(`
     CREATE OR REPLACE FUNCTION calculate_case_price(
-      plastic_units_per_case INT,
-      aluminium_units_per_case INT,
+      plastic_units_per_case NUMERIC,
+      aluminium_units_per_case NUMERIC,
       markup NUMERIC
     )
     RETURNS TABLE (
